Rename fetchGroups module to groups
